Avoid rake db:test:prepare for individual test runs
`rake db:test:prepare` is run when our test database is not in sync with
our schema. When using migrations, the test database is synced on running
`RAILS_ENV=test rake db:migrate`, but when we use `structure.sql` rails
has to run rake db:test:prepare before any test is run.

The patch
https://github.com/rails/rails/commit/61bacd6ef9c302238f3b8d9e6b432ee2090f6b78
merged into v4.2 introduced this behaviour. This was to fix the issue
https://github.com/rails/rails/issues/17170

This activerecord hack was proposed in response to an issue we had
created on github: https://github.com/rails/rails/issues/18982