Render projects/deleted when parent project is deleted
When a project is deleted and one tries to access any associated child
controller page, it results in a 500 error. We were using the scope
`from_param` that looks for non deleted projects. We get a nil value
which results in an UndefinedMethod error when `deleted?` is called on
the `nil` `@project` object.