Resolve TraceView conflict with Version
We were using the Version gem when we were considering using Docker for
our deployment.  However, Version defines a method 'is_versioned' on all
Modules. Unfortunately, the new TraceView gem was not getting that
method defined on it's base module, 'TraceViewBase'.

Since we aren't deploying with Docker and don't need the convenience
that the Version gem gave us, we dropping that gem.

Additionally, "Oboe" has been replaced with "TraceView", so the config
file has been globally updated