Fix tests after merging master into branch
One notes that we have a 'current_user' that is obtained from the
session[:account_id] value, which is an encrypted cookie set after
login.  This is different from the '@account' parameter used in things
like the AccountsController and PositionsController, which is the value
obtained from the 'id' or 'account_id' passed in as a URL paramter.

Note that the structure.sql comments out statements that are not
available in PostgreSQL 9.2 and the pg_repack library which one should
not expect all the development engineers to have on their system.