twitter_id and level change to Account Admin
This commit adds the twitter_id to the te Account Admin page so that
this value can be manually set for those accounts wherein the user is a
long-standing member of the Open Hub community and will not agree to use
Twitter Digits to authenticate.

Also in this commit is the change to the twitter_id validation to
enforce the uniqueness constraint only on account creation.  We will be
double checking that this change continues to contstrain users to only
one account.

Also in this commit is the addition of the "Admin" menu item to the user
profile menu