Fix the broken hooks_test test
The core issue was that the Ohloh Hamster account needs to be an
administrator.  Simply setting the level in the "create_" method wasn't
sufficient because *somewhere* there is a record inserted into the
accounts table with an ID 1 (this is used by other factories).

Really, I wasn't able to track it down.  I'm sure it's hidden in plain
sight.  As a solution, I just make sure that the found or created record
is set to the administrative level.  One notes that this account in
production is indeed an administrator.  And therefore, should not allow
anyone to log into it.