Interim checking - Changing approach
At this point, the polymorphic belongs_to behavior wherein a Job can
belong to a Project or a Repository was working.  From a project, one
could see all the jobs that belong to the Project and the Project's
repositories.  However, one could no longer simply access a job from the
dashboard.

Probably that's due to not having written any tests so there was no
regression coverage, otherwise I would have found this sooner.

It seems that the better approach will be to have the Job be a top level
entity without any of the associative relationships and to pass the id
of the parent entity -- project, repository, account, organization, etc.
-- as a command line parameter.  Not as elegant, but probably much more
straight forward.