Fix the definition of failed? for a Repository
There was a problem with the Ohloh implementation of failed? for a
Repository in that the order of a set of Jobs for a Repo was undefined,
and the code would take the first one and check to see if it had failed.
There was no guarantee that the first Job was the most recent.
We carried that forward.

And, we added a new defect in that we were not accounting for the
situation wherein all the jobs could have been completed.  In that case,
we had an empty array and "failed?" was called on the NilClass.

This change ensures ordering of the jobs and includes jobs in every
state.