Address nil job.current_step_at
When a Job is newly switched to the Running Status, there is a race
condition wherein the dashboard is loaded and fetches records from the
database and there is at least one Job in the Running state, but has a
null current_step_at field.

When this happens, the conditional logic to display the dashboard throws
an error.  The corrective action is to consider this state as a job that
has been running for less than five minutes.