Turn off cache_store for development
While testing the behavior of the defect described in OTWO-4096, it was
discovered that, even in the development environment, view code was
being cached.

This change enforces no caching in the development environment.
Alternatively, we can modify the .env file to use oh-stage-1 as the
default REDIS host (it is currently empty) and change the
.env.development file to be null.  This way we retain the configuration
information, but still have caching off by default in development.

OR, we can use the 'config.cache_store = nil' and simply comment out
that line when and if we want Rails caching in development.  Your
choice.