OTWO-4129  Add Optimizely tracking code
Marketing would like to track user behavior and demographic on the Open
Hub and has asked that we add this tracking code.

Note: Since this code is available on the public website for anyone
sufficiently clever enough to use "view source", there really isn't much
benefit from hiding this value in an environment variable.