Convert contributions view to materialized
Now that we are running Postgres 9.4 and the materialized view is able
of being updated with a concurrent flag, we can take advantage of using
the materialized view for the contributions view.  Requests for pages
that use this view typically take well over 30,000 ms for the query to
run.  Preliminary tests show over 50X improvement in query responses.

In order for the materialized view to be used, we need to call "refresh
materialized view" regularly, which will need to be done by a cron job.
Therefore, there is a SQL statement for regular execution included in
this commit.  In order to use the "refresh materialized view
concurrently", there needs to be a unique index on the view, which is
why the ID has an index.