Convert Contributions to Materialized View
The Contributions view is at the heart of just about all of our slowest
responding pages right now.  Pages accessing this view can regulary take
well over 1 to 2 seconds and more to process and generate. The
underlying reason is that whenever we access the Contributions view, the
underlying query is re-executed.  That's how views work.  Since the
Contributions view is already a join between some 4 tables, when we join
the Contributions view with other tables, we end up with some very large
and expensive queries.

Since we upgraded our production database to Postgres 9.4 we can take
advantage of MATERIALZED VIEWs, which were introduced into Postgres with
9.3. However, at that time, the view had to be locked in order to be
updated, which made them unsuitable for our uses.  With 9.4, the view
can be refreshed concurrently, which allows for continued reads on the
view while the update occurs.

This work drops the current view and replaces the same query with a
materialized view.  The real work is making sure the view is properly
refreshed.  Our first thoughts were that we would have a cron job to run
refresh on a regular basis.  However, while testing the change, it
became apparent that many aspects of our site will start failing unless
the view is updated at the correct point when there has been a related
data change.

At this point, all of the functionality seems to have been implemented.
However, code coverage is reporting that we are a bit under our target
test threshold. Additionally, the changes here to the models to call a
refresh on the view after certain events -- model creation, update,
deletes -- need to be reproduced in Ohloh since the work of generating
many of these model occurs there during analyses.