Add Call To Action for 2015 FOOS survey
The Future of Open Source (FOOS) survey is an annual survey conducted by
Black Duck.  Marketing has asked for the Open Hub to support the survey
by including a Call To Action to vote on the survey on the site. (After
injecting a butt-ugly header on top of all of our beautiful pages using
JS injection with their nasty optimizely JS library).

This change set includes the code for the masthead, styling and
restoration of the Optimizely JS library.

Additionally, the author was fed up with the Airbrake gem not ignoring
the fact that we are in the development environment and masking errors
when error could not be reported to Airbrake.  Ergo, the suppression of
reporting in the development and test environments.

Note that the FOOS survy, but NOT the Optimizely and Airbrake changes
need to be removed after April 1.