Address time issues for reliable results
The issue we were having with what ever it was that changed seems to
have been rooted in the way that times were being generated and handled
for Commit objects created by FactoryGirl as well as the time of the
Anaysis, as set by "Time.current".

When looking at the state of the data, it became apparent that the time
boundaries were lining up in a manner that prevented commits that had
appeared in the NamedCommits view from now appearing there.

The corrective action was to add some roughness to the time stamps. the
commits were created with specified times, instead of the default from
the factory, which was "Time.current.beginning_of_day"

Another issue is that the AnalysisSlocSet needs to have the as_of value
match the maximum position value of the commits in the code_set.