Remove invalid repos for GithubUser
A defect was found in production that a GitHub repository without any
code -- an empty repository -- will result in a GitRepository instance
being created but not persisted. Without the persistence, there is no
repository_id and the creation of an Enlistment fails.

This change ensures that only valid GitRepositories are added to the
array of repositories that are used to create the Enlistments.

However, the accompanying test does not ensure this functionality.
While the tests pass, removing the changed code in the GithubUser class
does not then cause the tests to fail.  Therefore the tests do not prove
the correctness of the fix.

The author grew frustrated with his efforts to jigger the tests so they
do the right thing while a real defect remains in production and has
decided to deploy the fix and back fill the test. Bad engineer.