Center "Review Security Info" button
Having removed the "Browse Code" button, the remaining "Review Security
Info" button needs to be displayed in the center of the Quick Reference
well so that the well remains balanced.

This change creates a full width row and divides it into 3 so that the
button in the middle will be balanced.