Add warning message to new registration page
We get a fair number of people who continue to created accounts that
violate the Terms Of Use.  Somewhere in the flow, we tell the user that
they have to comply with the TOU, but I can't remember the last time *I*
read one of those.

So, it seemed like a good idea to give our users some basic guidelines