Restore support for 'q' as well as 'query'
Code Center has been invoking Open Hub URLS with the '?q=some_string'
format consistently, but we seem to have dropped support for that with
Project PURR. The issue was recently raised. And while the Code Center
team can make a fix and push that into the new version of Code Center,
it is much easier and faster for the Open Hub to honor the 'q' query
parameter form as it used to.

This change set is focused on testing this string for the Project
controller although the change is low-level and should be available
every place we currently use "query" as a query string paramater.