Fix display of flash message in new project create
It looks like we were never able to display the flash message in the New
Project create action. A user was having difficulty adding their project
because they had the wrong branch, but there was no error. To them, it
seemed like the page simply refreshed when it was failing silently.

The fix is was to ensure the error flash was being populated.  Note that
this is a change from the use of the notice flash, which had been the
implemention.  However, it is the authors opinion that failure to save
should be indicated by an error message, instead of a notice.