Re-Replace logo and other stuff
To make peace, we agreed to use the new logo with the new Cyan and make
some small changes so the logo doesn't look so darn ugly with the rest
of the trade dress.

We changed the Cyan used in the main nav menu and highlight and the
"Discover, Track and Compare Open Source" box.  The author took the
opportunity to change the highlight of the Logged In User menu, but the
Twitter follow button is an image, so can't easily be changed and looks
freakin' wrong.  But the author was assured that everything would be
find and to not worry.