Rename migration so it matches fisbot
The original migration to rename the code_location_events table to
fisbot_events is in the fisbot code base.  By creating a new migration,
we essentially will have two different timestamps that are doing the
same thing, plus a less-than-elegant "if" statment that Rubocop didn't
like in the migration.

Therefore, this change set renames the migration and simplifies it.

Additionally, there are fixes to make the Rubocopy, Haml Lint and tests
pass