Refactor seed file
Based upon Priya's suggestion, I updated the seed file.  It's a little
easier on the eyes.  I looked into dynamically creating local variables
and assigning them to the values in the new "config" hash. But one
cannot create local variables any more. One could create local instance
variables, but that's overkill.  So we have the current for of the code.

Generating a string command makes it easy to verify that what will be
executed is what is intended.

And adding the port to the .env.test was necessary because no default
value is loaded. The default value is defined in .env.development, which
isn't invoked in non-development environemtns.