Initial checking:
Base models for LicensePermission and LicensePermissionRole

Using the "belongs_to :through" association so that there can be a role
for each license permission associated with a license.  A straight
has_and_belongs_to association would have only the FK ids, and there
wouldn't be a way to access the "status" field of the join table.

Therefore, there needs to be an model, LicensePermissionRole, that can
capture the relationship betweeen the License and the LicensePermission
and set the role