Models and relationships implemented
It is possible to create a LicensePermission and using the
License#license_permission_roles.create method to add the
LicensePermission to the License with the correct permission role --
permitted, forbidden, or required.

The License model has new filter methods for retrieving license
permissions by permission role:
License#permitted_license_permissions
License#forbidden_license_permissions
License#required_license_permissions

Next Step: Write a script to create the LicensePermissions and then add
them to the appropriate Licenses. After that, it will be possible to
complete the display of the licenses in the UI