OTWO-4995: Replace CTA with CoPilot
We are done with the 2017 Survey and need to replace the Call to Action.
However, when testing the change on Safari, the author noticed that the
search bar was not properly positioned.  Therefore, the home page
Billboard was reworked using the Bootstrap grid to ensure the page
displays the same.

Also note that the header banner has been removed from all non-home
pages.