OTWO-4996 Leverage Clearance for password validations
Clearance defines the following:
  validates :password, presence: true, unless: :skip_password_validation?
  ...
  def skip_password_validation?
    ... encrypted_password.present? && !encrypted_password_changed?
  end

This skip_password_validation? pattern ensures that the validation is
run only when creating account or updating account password. Using this
pattern helps us write more concise password validations.

We replaced skip_current_password_check with validate_current_password
to ensure that we check for current_password only when really need to.
The only place we really need it is in alter_passwords_controller.