Ignore TarballJobs when scheduling a fetch
The Code Location "schedule_fetch" would return if there were any
incomplete jobs, ostensibly because those jobs need to be completed
first.

However, the existence of a TarballJob causes problems with this. The
Project and Code Locations are not dependent upon a TarballJob
successfully completing. However, if there was a failed TarballJob, the
"incomplete" state is true, and no new job would be created.

The fix in this change set is to filter out TarballJobs from the set of
incomplete or new jobs (those that finished within the last 5 minutes).
Frankly, the author has some concerns about this blanket 5 minute
policy, but isn't attempting to address it now.

The test code illustrates that the TarballJob must belong to the Code
Location and have the Code Locations's best code set recoreded as well
because the "schedule_fetch" logic looks at the Code Set's jobs.

There is also a small rubocop fix to the Project License model.