OTWO-5039 Restore login with email
With the migration to Clearance, we lost the ability to login with an
email address.  This restores that feature.

There was also a regression failure where a disabled account was being
redirected to the "authenticate new" path. This was due to the
"!current_user.access.disabled?" in the
ApplicationController#redirect_for_spammer_verification method.  When a
user is disabled, then they'll get caught by a later before_action
filter, so this check isn't needed here.