refactored projects#show for OTWO colors, re 'REV 2'
This PR addresses the restyling of the OH projects#show page, using
colors that are addressed in the Synopsys Color Standards. Refer to
the mockup attached to the JIRA ticket for reference. Fyi, Some
sections on this page, ie: "Quick Reference", will have their
respective styling defined at a later date.