OTWO-5759: Update the Explore Projects page using the Mockup
This PR addresses the OTWO color updates to 'explore#projects',
and the embedded Highcharts demographics chart also on the page.

The version of Highcharts has been upgraded to 8.0.0 as to allow
the ability to style the charts with css (`styledMode = true`)
as opposed to defining all options within the code base, esp
with respect to colors.

This PR also updates a few other issues that were found while
researching Highcharts css features.