OTWO-5930: Update CSS for Account Show page
This PR addresses the updates re the new Ohloh UI Trade Dress to the
accounts#show page referenced in the 2019 OH Style Guide

This PR highlights the color updates to the commits_by_project chart,
as well as addressing some global color updates to such elements as
the timestamp signature, search box controls, and footer navigation
colors. These are also referenced in the 2019 OH Style Guide.