Cleanup unused test & development gems
binding_of_caller -> Binding.of_caller is not used anywhere.
guard -> If one needs guard, they can install it individually.
meta_request -> Used for a chrome extension
flog -> Install this locally if needed. We do not need it in Gemfile.
m -> Rails 5 already provides a way to run tests by line number.
ruby_parser -> we do not use this anywhere.
spring -> raises exceptions related to autoload on running single test.
capybara, selenium, jasmine & spinach -> No integration tests right now.
pry -> Install this locally if needed. We do not need it in Gemfile.
rails-erd -> Install this locally when needed.
rb-readline -> This is not used anywhere.
rubocop-faker -> We do not call rubocop with this extension.